HADOOP-19863. Pull count test to LocalFS
+ validate properly
